[datalink] move all hardcoded message parsing to modules datalink events
