+++ conf/modules/autopilot_gnc.xml
@@ -7,7 +7,7 @@
     </description>
   </doc>
   <dep>
-    <depends>system_core,@guidance,@navigation,@commands</depends>
+    <depends>system_core,@guidance,@navigation,@commands,autopilot_guided</depends>
     <provides>autopilot</provides>
   </dep>
   <header>
@@ -16,6 +16,7 @@
   <init fun="autopilot_init()"/>
   <periodic fun="autopilot_periodic()"/>
   <event fun="autopilot_event()"/>
+  <datalink message="GUIDED_SETPOINT_NED" fun="autopilot_guided_parse_GUIDED(buf)" cond="AP_MODE_GUIDED"/>
   <makefile target="!fbw">
     <file name="autopilot.c" dir="."/>
     <file name="autopilot_firmware.c" dir="$(SRC_FIRMWARE)"/>
@@ -23,9 +24,6 @@
     <file name="autopilot_static.c" dir="$(SRC_FIRMWARE)" cond="ifneq ($(USE_GENERATED_AUTOPILOT), TRUE)"/>
     <file name="autopilot_generated.c" dir="$(SRC_FIRMWARE)" cond="ifeq ($(USE_GENERATED_AUTOPILOT), TRUE)"/>
     <define name="USE_GENERATED_AUTOPILOT" value="1" cond="ifeq ($(USE_GENERATED_AUTOPILOT), TRUE)"/>
-  </makefile>
-  <makefile target="!fbw" firmware="rotorcraft">
-    <file name="autopilot_guided.c" dir="$(SRC_FIRMWARE)"/>
   </makefile>
 </module>
