+++ conf/modules/gps_ublox.xml
@@ -21,7 +21,8 @@
   <init fun="gps_ubx_init()"/>
   <periodic fun="gps_ubx_periodic_check()" freq="1." autorun="TRUE"/>
   <event fun="gps_ubx_event()"/>
-  <makefile target="ap|fbw">
+  <datalink message="HITL_UBX" fun="gps_ubx_parse_HITL_UBX(buf)"/> 
+  <makefile target="ap|fbw|hitl">
     <configure name="UBX_GPS_PORT" default="$(GPS_PORT)" case="upper|lower"/>
     <configure name="UBX_GPS_BAUD" default="$(GPS_BAUD)"/>
