+++ sw/airborne/firmwares/fixedwing/nav.c
@@ -584,6 +584,34 @@ void nav_without_gps(void)
 #endif
 }
 
+void nav_parse_BLOCK(struct link_device *dev, struct transport_tx *trans, uint8_t *buf)
+{
+  if (DL_BLOCK_ac_id(buf) != AC_ID) { return; }
+  nav_goto_block(DL_BLOCK_block_id(buf));
+  SEND_NAVIGATION(trans, dev);
+}
+
+void nav_parse_MOVE_WP(struct link_device *dev, struct transport_tx *trans, uint8_t *buf)
+{
+  if (DL_MOVE_WP_ac_id(buf) != AC_ID) { return; }
+  uint8_t wp_id = DL_MOVE_WP_wp_id(buf);
+
+  /* Computes from (lat, long) in the referenced UTM zone */
+  struct LlaCoor_f lla;
+  lla.lat = RadOfDeg((float)(DL_MOVE_WP_lat(buf) / 1e7f));
+  lla.lon = RadOfDeg((float)(DL_MOVE_WP_lon(buf) / 1e7f));
+  lla.alt = (float)(DL_MOVE_WP_alt(buf) / 1000.f); // mm to m
+  struct UtmCoor_f utm;
+  utm.zone = nav_utm_zone0;
+  utm_of_lla_f(&utm, &lla);
+  nav_move_waypoint(wp_id, utm.east, utm.north, utm.alt);
+
+  /* Waypoint range is limited. Computes the UTM pos back from the relative
+     coordinates */
+  utm.east = waypoints[wp_id].x + nav_utm_east0;
+  utm.north = waypoints[wp_id].y + nav_utm_north0;
+  pprz_msg_send_WP_MOVED(trans, dev, AC_ID, &wp_id, &utm.east, &utm.north, &utm.alt, &nav_utm_zone0);
+}
 
 /**************** 8 Navigation **********************************************/
