+++ sw/airborne/firmwares/rotorcraft/navigation.h
@@ -119,6 +119,8 @@ extern bool exception_flag[10];
 
 extern void nav_init(void);
 extern void nav_run(void);
+extern void nav_parse_BLOCK(uint8_t *buf);
+extern void nav_parse_MOVE_WP(uint8_t *buf);
 
 extern void set_exception_flag(uint8_t flag_num);
