+++ sw/airborne/firmwares/rover/navigation.h
@@ -162,6 +162,8 @@ extern uint8_t last_wp __attribute__((unused));
 
 extern void nav_init(void);
 extern void nav_run(void);
+extern void nav_parse_BLOCK(uint8_t *buf);
+extern void nav_parse_MOVE_WP(uint8_t *buf);
 
 extern void set_exception_flag(uint8_t flag_num);
