+++ sw/airborne/modules/datalink/datalink.c
@@ -57,126 +57,23 @@ void datalink_periodic(void)
   datalink_time++; // called at 1Hz
 }
 
+void datalink_parse_PING(struct link_device *dev, struct transport_tx *trans, uint8_t *buf)
+{
+  // Reply to the sender of the message
+  struct pprzlink_msg msg;
+  msg.trans = trans;
+  msg.dev = dev;
+  msg.sender_id = AC_ID;
+  msg.receiver_id = pprzlink_get_msg_sender_id(buf);
+  msg.component_id = 0;
+  pprzlink_msg_send_PONG(&msg);
+}
+
 void dl_parse_msg(struct link_device *dev, struct transport_tx *trans, uint8_t *buf)
 {
   uint8_t msg_id = pprzlink_get_msg_id(buf);
   uint8_t class_id = pprzlink_get_msg_class_id(buf);
-  uint8_t sender_id = pprzlink_get_msg_sender_id(buf);
-
-  // Check that the message is a datalink message
-  if (class_id == DL_datalink_CLASS_ID) {
-    /* parse datalink messages coming from ground station */
-    switch (msg_id) {
-      case  DL_PING:
-        {
-          // Reply to the sender of the message
-          struct pprzlink_msg msg;
-          msg.trans = trans;
-          msg.dev = dev;
-          msg.sender_id = AC_ID;
-          msg.receiver_id = sender_id;
-          msg.component_id = 0;
-          pprzlink_msg_send_PONG(&msg);
-        }
-        break;
-
-#ifndef INTERMCU_FBW
-      case DL_SETTING :
-        {
-          if (DL_SETTING_ac_id(buf) != AC_ID) { break; }
-          uint8_t i = DL_SETTING_index(buf);
-          float var = DL_SETTING_value(buf);
-          DlSetting(i, var);
-          // Reply to the sender of the message
-          struct pprzlink_msg msg;
-          msg.trans = trans;
-          msg.dev = dev;
-          msg.sender_id = AC_ID;
-          msg.receiver_id = sender_id;
-          msg.component_id = 0;
-          pprzlink_msg_send_DL_VALUE(&msg, &i, &var);
-        }
-        break;
-
-      case DL_GET_SETTING :
-        {
-          if (DL_GET_SETTING_ac_id(buf) != AC_ID) { break; }
-          uint8_t i = DL_GET_SETTING_index(buf);
-          float val = settings_get_value(i);
-          // Reply to the sender of the message
-          struct pprzlink_msg msg;
-          msg.trans = trans;
-          msg.dev = dev;
-          msg.sender_id = AC_ID;
-          msg.receiver_id = sender_id;
-          msg.component_id = 0;
-          pprzlink_msg_send_DL_VALUE(&msg, &i, &val);
-        }
-        break;
-#endif
-
-#ifdef RADIO_CONTROL_TYPE_DATALINK
-      case DL_RC_3CH :
-#ifdef RADIO_CONTROL_DATALINK_LED
-        LED_TOGGLE(RADIO_CONTROL_DATALINK_LED);
-#endif
-        parse_rc_3ch_datalink(
-            DL_RC_3CH_throttle_mode(buf),
-            DL_RC_3CH_roll(buf),
-            DL_RC_3CH_pitch(buf));
-        break;
-      case DL_RC_4CH :
-        if (DL_RC_4CH_ac_id(buf) == AC_ID) {
-#ifdef RADIO_CONTROL_DATALINK_LED
-          LED_TOGGLE(RADIO_CONTROL_DATALINK_LED);
-#endif
-          parse_rc_4ch_datalink(DL_RC_4CH_mode(buf),
-              DL_RC_4CH_throttle(buf),
-              DL_RC_4CH_roll(buf),
-              DL_RC_4CH_pitch(buf),
-              DL_RC_4CH_yaw(buf));
-        }
-        break;
-#endif // RADIO_CONTROL_TYPE_DATALINK
-
-#if USE_GPS
-      case DL_GPS_INJECT :
-        {
-          // Check if the GPS is for this AC
-          if (DL_GPS_INJECT_ac_id(buf) != AC_ID) { break; }
-
-          // GPS parse data
-          gps_inject_data(
-              DL_GPS_INJECT_packet_id(buf),
-              DL_GPS_INJECT_data_length(buf),
-              DL_GPS_INJECT_data(buf)
-              );
-        }
-        break;
-#if USE_GPS_UBX_RTCM
-      case DL_RTCM_INJECT :
-        {
-          // GPS parse data
-          gps_inject_data(DL_RTCM_INJECT_packet_id(buf),
-              DL_RTCM_INJECT_data_length(buf),
-              DL_RTCM_INJECT_data(buf));
-        }
-        break;
-#endif  // USE_GPS_UBX_RTCM
-#endif  // USE_GPS
-
-      default:
-        break;
-    }
-  }
-  /* Parse firmware specific datalink */
-  firmware_parse_msg(dev, trans, buf);
-
   /* Parse modules datalink */
   modules_parse_datalink(msg_id, class_id, dev, trans, buf);
 }
 
-/* default empty WEAK implementation for firmwares without an extra firmware_parse_msg */
-WEAK void firmware_parse_msg(struct link_device *dev __attribute__((unused)), struct transport_tx *trans __attribute__((unused)), uint8_t *buf __attribute__((unused)))
-{
-}
